Enemy Update
Created the Damage function and for the Player and also made the death animations for the zombie. The enemy can now receive damage form all weapons.  I am planning to add headshot dmg and i have to add sounds for the enemy. Last but not least I have to create the enemy's attack dmg function.